added class databes. Changes in Picker and Main activities

+++ app/src/main/java/com/example/schedule/model/MyJSONFile.kt
@@ -2,5 +2,5 @@ package com.example.schedule.model
 
 data class MyJSONFile(
     var pairs: List<PairClass>,
-    var version: Double
+    var version: Long
 )
